Projeção de pizzas com metamodel

+++ src/main/java/com/uds/pizzaria/model/Adicional.java
@@ -19,7 +19,7 @@ public class Adicional {
     private Long id;
 
     private String descricao;
-    private int valor;
-    private int tempo;
+    private Long valor;
+    private Long tempo;
 
 }
